Update diffable target code for SPM
